updating forms with formElement and submit event for Login
