+++ src/utils/inputValidators/loginValidator.ts
@@ -1,6 +1,7 @@
 export default function loginValidator(e: { target: { value: string; }; }) {
-  const inputValue = e.target.value.trim(); // Remove leading and trailing whitespaces
-  const isValid = /^[a-zA-Zа-яА-Я0-9_-]{3,20}$/.test(inputValue); // Regular expression for validation
+  const inputValue = e.target.value.trim(); 
+  const isValid = /^[a-zA-Zа-яА-Я0-9_-]{3,20}$/.test(inputValue); 
+
 
   if (isValid) {
     this.children.LoginInput.setProps({ error: false, inputText: null });
@@ -9,5 +10,6 @@ export default function loginValidator(e: { target: { value: string; }; }) {
   }
 
   this.setProps({login: inputValue});
+
 
 }
